+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cli;
 pub mod cluster;
+pub mod cluster_rewrite;
 pub mod error;
 pub mod evil;
 pub mod logging;